register.py: Do not chown token file in container
Inside a container, chowning a token to the condor user generally
creates a token that is not accessible by the user who ran the
container.

This is especially true if user namespaces are used (e.g. with podman),
since it will be chowned to one of the users' subuids, and they would
have to `sudo` or `podman unshare` to chown it to something they can
use.